serverless_functions/image_classification/requirements.txt
google-cloud-firestore
google-cloud-vision